# strings: smarter growth of temporal buffer and avoid copying on return